gather_pre_rms_norm除零 customize_blockcopy整数溢出风险修复

Created-by: zhuzemao
Commit-by: zhuzemao
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## checklist
<!--  [x] 表示选中 -->
- [x] 是否通过本地IDE对代码进行静态检查
- [x] 是否通过本地IDE对代码进行格式化处理
- [x] 是否进行空指针校验
- [x] 是否进行返回值校验
- [x] 是否正确释放new/malloc申请的内存
- [x] 是否充分考虑接口的异常场景
- [x] 是否正确记录错误日志

See merge request: cann/ascend-transformer-boost!1703

+++ ops_customize/ops/customize_blockcopy/kernel_implement/op_kernel/customize_blockcopy.cpp
@@ -96,7 +96,7 @@ private:
             gmOffset_ = perCoreCopyCount_ * blockIdx_ + tailCoreCopyCount_;
         }
         blockCount_ = tiling_data->blockCount;
-        blockSizeinElement_ = tiling_data->blockSize * tiling_data->numHead * tiling_data->headSizeK;
+        blockSizeinElement_ = static_cast<uint64_t>(tiling_data->blockSize) * tiling_data->numHead * tiling_data->headSizeK;
         sourceCount_ = tiling_data->sourceCount;
         destinationCount_ = tiling_data->destinationCount;
         typeByte_ = tiling_data->typeByte;
